Separate signal code from service code
* Preparation for improving ignoring signals.
