+++ mapper/cumulocity/c8y_mapper/src/main.rs
@@ -4,6 +4,7 @@ use service::*;
 
 mod mapper;
 mod service;
+mod signals;
 
 const DEFAULT_LOG_LEVEL: &str = "warn";
