+++ mapper/cumulocity/c8y_mapper/src/signals/mod.rs
@@ -1,0 +1,11 @@
+#[cfg(windows)]
+pub use windows::*;
+
+#[cfg(not(windows))]
+pub use unix::*;
+
+#[cfg(not(windows))]
+mod unix;
+
+#[cfg(windows)]
+mod windows;
