+++ mapper/cumulocity/c8y_mapper/src/signals/unix.rs
@@ -1,0 +1,13 @@
+use tokio::signal::unix::{signal, Signal, SignalKind};
+
+pub fn sighup_stream() -> std::io::Result<Signal> {
+    signal(SignalKind::hangup())
+}
+
+pub fn sigterm_stream() -> std::io::Result<Signal> {
+    signal(SignalKind::terminate())
+}
+
+pub fn sigint_stream() -> std::io::Result<Signal> {
+    signal(SignalKind::interrupt())
+}
